LayoutTests/imported/w3c/web-platform-tests/referrer-policy/gen/srcdoc-inherit.meta/origin-when-crossorigin/worker-module.http-expected.txt

CONSOLE MESSAGE: Failed to set referrer policy: The value 'origin-when-crossorigin' is not one of 'no-referrer', 'no-referrer-when-downgrade', 'same-origin', 'origin', 'strict-origin', 'origin-when-cross-origin', 'strict-origin-when-cross-origin' or 'unsafe-url'.

PASS Referrer Policy: Expects stripped-referrer for worker-module to same-http origin and keep-origin redirection from http context.
PASS Referrer Policy: Expects stripped-referrer for worker-module to same-http origin and no-redirect redirection from http context.
